requirements.txt
diffusers
gradio
Pillow
qrcode
torch
